Inherit minutes per period.

+++ src/main/java/net/sf/mpxj/primavera/PrimaveraDatabaseReader.java
@@ -214,8 +214,8 @@ public final class PrimaveraDatabaseReader extends AbstractProjectReader
          ProjectProperties ph = m_reader.getProject().getProjectProperties();
          ph.setCreationDate(row.getDate("create_date"));
          ph.setLastSaved(row.getDate("update_date"));
-         ph.setMinutesPerDay(Double.valueOf(row.getDouble("day_hr_cnt").doubleValue() * 60));
-         ph.setMinutesPerWeek(Double.valueOf(row.getDouble("week_hr_cnt").doubleValue() * 60));
+         ph.setMinutesPerDay(Integer.valueOf((int)(row.getDouble("day_hr_cnt").doubleValue() * 60)));
+         ph.setMinutesPerWeek(Integer.valueOf((int)(row.getDouble("week_hr_cnt").doubleValue() * 60)));
          ph.setWeekStartDay(Day.getInstance(row.getInt("week_start_day_num")));
 
          processDefaultCurrency(row.getInteger("curr_id"));
